+++ include/saucer/window.hpp
@@ -58,16 +58,16 @@ namespace saucer
 
       public:
         [[thread_safe]] void clear(window_event event);
-        [[thread_safe]] void remove(window_event event, std::size_t id);
+        [[thread_safe]] void remove(window_event event, std::uint64_t id);
 
         template <window_event Event> //
-        [[thread_safe]] std::size_t on(events::callback_t<Event> &&) = delete;
+        [[thread_safe]] std::uint64_t on(events::callback_t<Event> &&) = delete;
 
       public:
         static void run();
     };
 #include "annotations.hpp" //NOLINT
 
-    template <> std::size_t window::on<window_event::close>(events::callback_t<window_event::close> &&);
-    template <> std::size_t window::on<window_event::resize>(events::callback_t<window_event::resize> &&);
+    template <> std::uint64_t window::on<window_event::close>(events::callback_t<window_event::close> &&);
+    template <> std::uint64_t window::on<window_event::resize>(events::callback_t<window_event::resize> &&);
 } // namespace saucer
